+++ lib/zenlish/wclasses/distributive_determiner.rb
@@ -1,0 +1,11 @@
+require_relative 'determiner'
+
+module Zenlish
+  module WClasses
+    # Distributive determiner (aka distributive adjective) is a class of determiner,
+    # that is used to refer to individual members within a group or within a pair.
+    # Examples: each, every, either, neither.
+    class DistributiveDeterminer < Determiner
+    end # class
+  end # module
+end # module