+++ lib/zenlish/wclasses/fronting_quantifier.rb
@@ -1,0 +1,9 @@
+require_relative 'quantifier'
+
+module Zenlish
+  module WClasses
+    # Quantifiers that appear in front of determiners: all, both
+    class  FrontingQuantifier < Quantifier
+    end # class
+  end # module
+end # module