+++ lib/zenlish/wclasses/quantifier.rb
@@ -1,0 +1,10 @@
+require_relative 'determiner'
+
+module Zenlish
+  module WClasses
+    # Quantifiers are used to indicate the number or quantity of the noun under consideration.
+    # enough, few, little, many, more, most, much, none, several, some
+    class  Quantifier < Determiner
+    end # class
+  end # module
+end # module